Refactoring packages to fix checkstyles warnings
