+++ src/main/java/seedu/address/logic/commands/item/IncreaseItemCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.ItemCommand;
+package seedu.address.logic.commands.item;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ITEM_QUANTITY;
@@ -31,7 +31,7 @@ public class IncreaseItemCommand extends Command {
             + "Example: " + COMMAND_WORD + " 1 "
             + PREFIX_ITEM_QUANTITY + "4";
 
-    public static final String MESSAGE_INCREASE_ITEM_SUCCESS = "Increased quantity of Item: %1$s";
+    public static final String MESSAGE_INCREASE_ITEM_SUCCESS = "Increased quantity of item: %1$s";
 
     private final Index index;
     private final ItemQuantity quantityToAdd;
