+++ src/main/java/seedu/address/logic/parser/item/DecreaseItemCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser.Item;
+package seedu.address.logic.parser.item;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
@@ -7,8 +7,8 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_ITEM_QUANTITY;
 import java.util.stream.Stream;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.ItemCommand.DecreaseItemCommand;
-import seedu.address.logic.commands.ItemCommand.IncreaseItemCommand;
+import seedu.address.logic.commands.item.DecreaseItemCommand;
+import seedu.address.logic.commands.item.IncreaseItemCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
