+++ src/main/java/seedu/address/logic/parser/item/DeleteItemCommandParser.java
@@ -1,9 +1,9 @@
-package seedu.address.logic.parser.Item;
+package seedu.address.logic.parser.item;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.ItemCommand.DeleteItemCommand;
+import seedu.address.logic.commands.item.DeleteItemCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
