+++ src/main/java/seedu/address/model/item/ItemLocation.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
 
 /**
- *  Represents a Item's location in the inventory.
+ *  Represents a item's location in the inventory.
  *  Guarantees: immutable; is valid as declared in {@link #isValidItemLocation(String)}
  */
 public class ItemLocation {
