+++ src/main/java/seedu/address/model/item/ItemNameContainsKeywordsPredicate.java
@@ -6,7 +6,7 @@ import java.util.function.Predicate;
 import seedu.address.commons.util.StringUtil;
 
 /**
- * Tests that a {@code Item}'s {@code ItemName} matches any of the keywords given.
+ * Tests that a {@code item}'s {@code ItemName} matches any of the keywords given.
  */
 public class ItemNameContainsKeywordsPredicate implements Predicate<Item> {
     private final List<String> keywords;
